refactor: extract colorimetry-io into standalone spectral-io crate
The io/ crate is superseded by the new standalone spectral-io crate at
https://github.com/harbik/spectral-io. This commit wires the decoupled
crate back into colorimetry as an optional dependency.

- Remove io/ workspace member and all its files
- Add spectral-io as optional dependency (feature "spectral-io")
- Add [patch.crates-io] entry pointing to ../spectral-io for local dev
- Add src/spectrum/from_spectral_io.rs: IntoSpectrum extension trait
  implementing to_spectrum_{linear,sprague,smooth,binned} for
  spectral_io::SpectrumRecord using colorimetry::Spectrum as output
- Re-export IntoSpectrum from crate root under feature "spectral-io"
