--- io/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "colorimetry-io"
-version = "0.1.0"
-edition = "2021"
-description = "Reader and validator for UV-Vis / visible spectral data JSON files"
-license = "MIT"
-readme = "README.md"
-
-[features]
-default = ["colorimetry", "spectrashop"]
-colorimetry = ["dep:colorimetry"]
-spectrashop = []
-
-[[example]]
-name = "spectrashop_to_json"
-required-features = ["spectrashop"]
-
-[dependencies]
-colorimetry = { workspace = true, optional = true }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-thiserror = "1"
