--- io/examples/black_ace_licorice.txt
@@ -1,63 +1,0 @@
-SpectraShop	5.0
-NUMBER_OF_SETS	1
-ORIGINATOR	"Robin D. Myers"
-CREATED	"2007-09-15"
-MANUFACTURER	"Food Savvy LLC"
-PROD_DATE	"2007"
-SPECTRUM_TYPE	"Reflective"
-MATERIAL	"Licorice"
-SURFACE	"Matte"
-INSTRUMENTATION	"Spectrolino"
-INSTRUMENT_SERIAL_NUMBER	""
-MEASUREMENT_GEOMETRY	"45/0"
-MEASUREMENT_SOURCE	"D65"
-MEASUREMENT_APERTURE	"4 mm"
-MEASUREMENT_FILTER	"D65"
-SAMPLE_BACKING	"Self"
-NMEASUREMENTS	9
-OBSERVER	"2 degree"
-ILLUMINANT	"D65"
-RGB_SPACE	"sRGB"
-NUMBER_OF_FIELDS	75
-BEGIN_DATA_FORMAT
-SAMPLE_ID1	SAMPLE_ID2	SAMPLE_ID3	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL	SPECTRAL_NM	SPECTRAL_VAL
-END_DATA_FORMAT
-BEGIN_DATA
-"Black Ace Licorice"	""	""	
-380	4.588889E-2
-390	4.773333E-2
-400	4.816667E-2
-410	4.898889E-2
-420	4.900000E-2
-430	4.936666E-2
-440	4.961111E-2
-450	4.978889E-2
-460	4.983333E-2
-470	5.001111E-2
-480	5.003333E-2
-490	5.014444E-2
-500	5.036667E-2
-510	5.042222E-2
-520	5.080000E-2
-530	5.067778E-2
-540	5.022222E-2
-550	5.015555E-2
-560	5.054444E-2
-570	5.153333E-2
-580	5.115555E-2
-590	5.043333E-2
-600	5.031111E-2
-610	5.093333E-2
-620	5.148889E-2
-630	5.161111E-2
-640	5.125555E-2
-650	5.154444E-2
-660	5.218889E-2
-670	5.354444E-2
-680	5.532222E-2
-690	5.702222E-2
-700	5.932222E-2
-710	6.144444E-2
-720	6.401110E-2
-730	6.696666E-2
-END_DATA
